// src/restaurant/interfaces/searched-menu-by-slug.ts
import { Image } from '@/shared/interfaces';
import { NutritionalInformation, Rating } from '.';

// Generated by https://quicktype.io

export interface SearchedMenuBySlug {
  _id: string;
  name: string;
  description: string;
  price: number;
  discount: number;
  stock: number;
  stockDaily: number;
  category: CategorySearchedMenuBySlug;
  restaurant: RestaurantSearchedMenuBySlug;
  isActive: boolean;
  nutritionalInformation: NutritionalInformation[];
  ingredients: string[];
  images: Image[];
  ratings: Rating[];
  createdAt: string;
  updatedAt: string;
  slug: string;
  __v: number;
}

export interface CategorySearchedMenuBySlug {
  _id: string;
  name: string;
  banner: Image;
  image: Image;
  slug: string;
}

export interface RestaurantSearchedMenuBySlug {
  _id: string;
  name: string;
  address: string;
  logo: Image;
  slug: string;
}